start: allow overriding jira project ids
I wanted to say `start -p whatever REP-1234`, where REP-1234 would have
gotten the "misc" project. That didn't work, and it wound up with
"misc"; this commit makes it be in the project "whatever," as I wanted.

This is sort of a goofy way to do it, because go's zero values are kind
of goofy, but you dance with the one that brung you.